Introduce comments
We are using the consolidated `django-contrib-comments` here.
It used to be called `django.contrib.comments` but was split into
a separate package.

We've had to hack around some of it's features, tho.

First, we only accept comments from authenticated users. That is probably not
enforced at all places right now, but we've decorated the original
`post_comment` view in our `comments2` application.

We also removed the `comment` field from all objects. Commentable objects have
to put a comment field in the form, get it's cleaned data and create a comment
through `comments2.utils.create_comment`.