Little tweaks to registration process
* Redirect the user back to home after activation. With an info message.

* Display activation message within our base template.

This is not ideal, tho. We should redirect the user "back to where he came
from" after registration is complete. But we can't do that now, because we
have no reliable way to show flash messages to the user (since our flash
messages currently suck and don't even work).

In order to accomplish this, tho, we would have to implement the `next` field
in the registration view, then save it in the registration profile, then
redirect the user if he come quick enough. It's a real mess and I can't deal
with it right now.