Support sign in through social networks.
Here we add support to Google OAuth 2, Facebook Login and Twitter OAuth.

We used `python-social-auth` to do the hard work. There were some minor
needed tweaks in the pipeline because we don't have usernames, just user
emails.

Also, the icons for the buttons are not included. Someone should do it in the
next commits.